feat: add copy and download btns to the operation-index
