+++ src/components/viewer/OperationViewer.tsx
@@ -22,6 +22,7 @@ import { useAtom } from 'jotai'
 import { noop } from 'lodash-es'
 import { ComponentType, FC, useEffect, useMemo, useState } from 'react'
 import { Link } from 'react-router-dom'
+import { handleCopyShortCode, handleDownloadJSON } from 'services/operation'
 
 import { FactItem } from 'components/FactItem'
 import { Paragraphs } from 'components/Paragraphs'
@@ -141,39 +142,6 @@ export const OperationViewer: ComponentType<{
       }
     }, [error])
 
-    const handleCopyShortCode = () => {
-      const shortCode = toShortCode(operation.id)
-      navigator.clipboard.writeText(shortCode)
-
-      AppToaster.show({
-        message: '已复制神秘代码，前往 MAA 粘贴即可使用~',
-        intent: 'success',
-      })
-    }
-
-    const handleDownloadJSON = () => {
-      // pretty print the JSON
-      const json = JSON.stringify(
-        snakeCaseKeysUnicode(operationDoc, { deep: true }),
-        null,
-        2,
-      )
-      const blob = new Blob([json], {
-        type: 'application/json',
-      })
-      const url = URL.createObjectURL(blob)
-      const link = document.createElement('a')
-      link.href = url
-      link.download = `MAACopilot_${operationDoc.doc.title}.json`
-      link.click()
-      URL.revokeObjectURL(url)
-
-      AppToaster.show({
-        message: '已下载作业 JSON 文件，前往 MAA 选择即可使用~',
-        intent: 'success',
-      })
-    }
-
     const handleRating = async (decision: OpRatingType) => {
       // cancel rating if already rated by the same type
       if (decision === operation.ratingType) {
@@ -220,7 +188,7 @@ export const OperationViewer: ComponentType<{
               className="ml-4"
               icon="download"
               text="下载原 JSON"
-              onClick={handleDownloadJSON}
+              onClick={() => handleDownloadJSON(operationDoc)}
             />
 
             <Button
@@ -228,7 +196,7 @@ export const OperationViewer: ComponentType<{
               icon="clipboard"
               text="复制神秘代码"
               intent="primary"
-              onClick={handleCopyShortCode}
+              onClick={() => handleCopyShortCode(operation)}
             />
           </>
         }
